// playwright-e2e/enums/user-key.ts
enum UserKey {
  CLAIMANT_SOLICITOR = 'claimant-solicitor',
  CLAIMANT_ORGANISATION_SUPER = 'claimant-organisation-super',
  CLAIMANT_SOLICITOR_BULK_SCAN = 'claimant-solicitor-bulk-scan',
  DEFENDANT_SOLICITOR_1 = 'defendant-solicitor-1',
  DEFENDANT_SOLICITOR_2 = 'defendant-solicitor-2',
  CIVIL_ADMIN = 'civil-admin',
  NBC_REGION_1 = 'nbc-region-1',
  NBC_REGION_2 = 'nbc-region-2',
  NBC_REGION_4 = 'nbc-region-4',
  NBC_LOCAL = 'nbc-local',
  JUDGE_REGION_1 = 'judge-region-1',
  JUDGE_REGION_2 = 'judge-region-2',
  JUDGE_REGION_4 = 'judge-region-4',
  HEARING_CENTER_ADMIN_LOCAL = 'hearing-admin-local',
  HEARING_CENTER_ADMIN_REGION_1 = 'hearing-center-admin-region-1',
  HEARING_CENTER_ADMIN_REGION_2 = 'hearing-center-admin-region-2',
  HEARING_CENTER_ADMIN_REGION_4 = 'hearing-center-admin-region-4',
  TRIBUNAL_CASEWORKER_REGION_4 = 'tribunal-caseworker-region-4',
  CTSC_ADMIN = 'ctsc-admin',
  CIVIL_SYSTEM_UPDATE = 'system_update',
}

export default UserKey;